[Fix] Implicit conversions to unsigned integers

+++ tests/src/io/utils.cpp
@@ -161,7 +161,7 @@ void testIOMethods() {
             content);
 
   // C Read
-  size_t charCount = -1;
+  size_t charCount = static_cast<size_t>(-1);
   char *c = occa::io::c_read(test_foo, &charCount);
   ASSERT_EQ(charCount,
             content.size());
